refactor(movie): service 구현 수정
