+++ src/main/java/com/movieproject/domain/movie/controller/MovieController.java
@@ -5,7 +5,6 @@ import com.movieproject.common.response.PageResponse;
 import com.movieproject.domain.movie.dto.response.MovieListDto;
 import com.movieproject.domain.movie.dto.request.MovieRequestDto;
 import com.movieproject.domain.movie.dto.response.MovieResponseDto;
-import com.movieproject.domain.movie.service.MovieExternalService;
 import com.movieproject.domain.movie.service.MovieInternalService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -20,27 +19,24 @@ import org.springframework.web.bind.annotation.*;
 public class MovieController {
 
     private final MovieInternalService movieInternalService;
-    private final MovieExternalService movieExternalService;
 
     @PostMapping
     public ResponseEntity<ApiResponse<MovieResponseDto>> createMovie(
             @Valid @RequestBody MovieRequestDto.Create requestDto) {
-
         MovieResponseDto responseDto = movieInternalService.createMovie(requestDto);
-
         return ApiResponse.created(responseDto, "영화가 성공적으로 등록되었습니다.");
-    }
-
-    @GetMapping
-    public ResponseEntity<ApiResponse<PageResponse<MovieListDto>>> getMovies(Pageable pageable) {
-        Page<MovieListDto> moviesPage = movieExternalService.getMovies(pageable);
-        return ApiResponse.pageSuccess(moviesPage, "영화 목록 조회가 완료되었습니다.");
     }
 
     @GetMapping("/{movieId}")
     public ResponseEntity<ApiResponse<MovieResponseDto>> getMovieInfo(
             @PathVariable Long movieId) {
-        MovieResponseDto responseDto = movieExternalService.getMovieInfo(movieId);
+        MovieResponseDto responseDto = movieInternalService.getMovieInfo(movieId);
         return ApiResponse.success(responseDto, "영화 상세 정보 조회가 완료되었습니다.");
+    }
+
+    @GetMapping
+    public ResponseEntity<ApiResponse<PageResponse<MovieListDto>>> getMovies(Pageable pageable) {
+        Page<MovieListDto> moviesPage = movieInternalService.getMovies(pageable);
+        return ApiResponse.pageSuccess(moviesPage, "영화 목록 조회가 완료되었습니다.");
     }
 }
