+++ src/test/java/com/movieproject/movie/MovieInternalServiceTest.java
@@ -4,6 +4,7 @@ import com.movieproject.domain.director.entity.Director;
 import com.movieproject.domain.director.service.DirectorExternalService;
 import com.movieproject.domain.movie.dto.request.MovieRequestDto;
 import com.movieproject.domain.movie.dto.response.MovieResponseDto;
+import com.movieproject.domain.movie.dto.response.MovieListDto;
 import com.movieproject.domain.movie.entity.Movie;
 import com.movieproject.domain.movie.repository.MovieRepository;
 import com.movieproject.domain.movie.service.MovieInternalService;
@@ -13,15 +14,19 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.mockito.internal.verification.VerificationModeFactory.times;
 
 @ExtendWith(MockitoExtension.class)
 class MovieInternalServiceTest {
@@ -40,18 +45,12 @@ class MovieInternalServiceTest {
     void createMovie_success() {
         // given
         Long directorId = 1L;
-        Long movieId = 100L;
         MovieRequestDto.Create requestDto = new MovieRequestDto.Create(
                 "테스트 영화", LocalDate.of(2025, 1, 1), 120, "KOREA", "테스트 장르", directorId
         );
-
-        Director testDirector = Director.of("테스트 감독", "KOREA", LocalDate.of(2024, 1, 1));
-        ReflectionTestUtils.setField(testDirector, "directorId", directorId);
-
+        Director testDirector = Director.of("테스트 감독", "KOREA", LocalDate.now());
         Movie mockMovie = Movie.builder().director(testDirector).title(requestDto.title()).build();
-        ReflectionTestUtils.setField(mockMovie, "id", movieId);
 
-        // Mockito 동작 정의
         when(directorExternalService.findDirectorById(directorId)).thenReturn(testDirector);
         when(movieRepository.save(any(Movie.class))).thenReturn(mockMovie);
 
@@ -60,9 +59,46 @@ class MovieInternalServiceTest {
 
         // then
         assertThat(responseDto).isNotNull();
+        assertThat(responseDto.title()).isEqualTo("테스트 영화");
+    }
+
+    @Test
+    @DisplayName("영화 상세 정보 조회 성공 테스트")
+    void getMovieInfo_success() {
+        // given
+        Long movieId = 1L;
+        Director testDirector = Director.of("테스트 감독", "KOREA", LocalDate.now());
+        Movie mockMovie = Movie.builder().title("테스트 영화").director(testDirector).build();
+        ReflectionTestUtils.setField(mockMovie, "id", movieId);
+
+        when(movieRepository.findById(movieId)).thenReturn(Optional.of(mockMovie));
+
+        // when
+        MovieResponseDto responseDto = movieInternalService.getMovieInfo(movieId);
+
+        // then
+        assertThat(responseDto).isNotNull();
         assertThat(responseDto.movieId()).isEqualTo(movieId);
         assertThat(responseDto.title()).isEqualTo("테스트 영화");
-        verify(directorExternalService, times(1)).findDirectorById(directorId);
-        verify(movieRepository, times(1)).save(any(Movie.class));
+    }
+
+    @Test
+    @DisplayName("영화 목록 조회 (페이징) 성공 테스트")
+    void getMovies_success() {
+        // given
+        Movie movie1 = Movie.builder().title("영화1").build();
+        Movie movie2 = Movie.builder().title("영화2").build();
+        Pageable pageable = PageRequest.of(0, 10);
+        Page<Movie> mockPage = new PageImpl<>(List.of(movie1, movie2), pageable, 2);
+
+        when(movieRepository.findAll(pageable)).thenReturn(mockPage);
+
+        // when
+        Page<MovieListDto> resultPage = movieInternalService.getMovies(pageable);
+
+        // then
+        assertThat(resultPage).isNotNull();
+        assertThat(resultPage.getTotalElements()).isEqualTo(2);
+        assertThat(resultPage.getContent().get(0).title()).isEqualTo("영화1");
     }
 }
